backend/requirements_new.txt
fastapi==0.135.3
motor==3.3.1
pydantic==2.12.5
PyJWT==2.12.1
pymongo==4.5.0
python-dotenv==1.2.2
python_bcrypt==0.3.2
Requests==2.33.1
starlette==1.0.0
